Add simple window creating

+++ Source/Engine/EngineConfig.hpp
@@ -15,8 +15,8 @@ struct EngineConfig
 {
     bool isEditMode;
     bool isFullscreen;
-    uint32_t windowWidth;
-    uint32_t windowHeight;
+    uint16_t windowWidth;
+    uint16_t windowHeight;
     bool isMaximized;
 };
